Allow switching out the default shaders

+++ src/default_vertex_shader.glsl
@@ -1,7 +1,7 @@
 #version 330 core
 
-layout(location = 0) in ivec2 pos;
-layout(location = 1) in ivec2 uv;
+layout(location = 0) in vec2 pos;
+layout(location = 1) in vec2 uv;
 
 out vec2 v_uv;
